playground/requirements.txt
langchain==0.0.123
openai==0.27.2
python-dotenv==1.0.0
pandas==1.5.3